refactor: SelectedCountriesContext into its own folder
