+++ src/view/provider/RootProvider.tsx
@@ -1,20 +1,11 @@
-import React, { createContext } from "react";
-import {
-    SelectedCountriesContextType,
-    useSelectedCountriesContext,
-} from "./useSelectedCountriesContext";
+import React from "react";
+import { SelectedCountriesProvider } from "./selected-countries/SelectedCountriesProvider";
 interface RootProviderProps {
     children: React.ReactNode;
 }
-export const SelectedCountriesContext =
-    createContext<SelectedCountriesContextType | null>(null);
 
 export const RootProvider = (props: RootProviderProps) => {
-    const selectedCountriesContext = useSelectedCountriesContext();
-
     return (
-        <SelectedCountriesContext.Provider value={selectedCountriesContext}>
-            {props.children}{" "}
-        </SelectedCountriesContext.Provider>
+        <SelectedCountriesProvider>{props.children}</SelectedCountriesProvider>
     );
 };
